deuxiéme-commit

+++ app/Http/Controllers/AdminController.php
@@ -4,9 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\ServiceRequeste;
 use App\Http\Requests\UserRequest;
+use App\Http\Resources\UserServiceResource;
 use App\Models\Service;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
 
 class AdminController extends Controller
@@ -18,7 +20,19 @@ class AdminController extends Controller
     {
         $users = User::all();
         return response()->json([
-            "users" => $users
+            "data" => [
+                "users" => $users
+            ]
+        ]);
+    }
+
+    public function listeService()
+    {
+        $services = Service::all();
+        return response()->json([
+            "data" => [
+                "services" => UserServiceResource::collection($services),
+            ]
         ]);
     }
 
@@ -31,9 +45,11 @@ class AdminController extends Controller
             "nom" => $request
         ]);
         return response()->json([
-            "status" => 200,
-            "message" => "Isertion resuisie",
-            "user" => $user
+            "data" => [
+                "status" => 200,
+                "message" => "Isertion resuisie",
+                "user" => $user
+            ]
         ]);
     }
 
@@ -42,20 +58,42 @@ class AdminController extends Controller
      */
     public function store(UserRequest $request)
     {
+        return DB::transaction(function()use($request){
+            $service = Service::find($request->service_id);
+               // Vérifier si l'utilisateur avec le même email existe déjà
+        $existingEmail = User::where('email', $request->email)->first();
+        $existingTelephone = User::where('telephone', $request->telephone)->first();
+        if ($existingEmail || $existingTelephone) {
+            // Utilisateur avec le même email existe déjà, retournez une réponse d'erreur
+            return response()->json([
+                "data" => [
+                    "status" => 400,
+                    "message" => "L'utilisateur avec cet email existe déjà."
+                ]
+            ], 400);
+        }
+        // L'utilisateur n'existe pas, donc nous le créons
         $user = User::create([
             "nom" => $request->nom,
             "prenom" => $request->prenom,
             "telephone" => $request->telephone,
             "email" => $request->email,
-            "password" => Hash::make($request->password),
+            "password" => Hash::make(12345),
             "role" => $request->role,
+            "specialite" => $request->specialite,
         ]);
+        $user->services()->attach($service);
         return response()->json([
-            "status" => 200,
-            "message" => "Isertion resuisie",
-            "user" => $user
+            "data" => [
+                "status" => 200,
+                "message" => "Insertion réussie",
+                "user" => $user
+            ]
         ]);
+        });
+
     }
+
     /**
      * Display the specified resource.
      */
@@ -63,8 +101,10 @@ class AdminController extends Controller
     {
         $user = User::select('users.*')->where('id', $id)->first();
         return response()->json([
-            "status" => 200,
-            "user" => $user,
+            "data" => [
+                "status" => 200,
+                "user" => $user,
+            ]
         ]);
     }
 
@@ -82,19 +122,41 @@ class AdminController extends Controller
     public function update(Request $request, string $id)
     {
         $user = User::find($id);
-        $user->update([
-            "nom" => $request->nom,
-            "prenom" => $request->prenom,
-            "telephone" => $request->telephone,
-            "email" => $request->email,
-            "password" => bcrypt($request->password),
-            "role" => $request->role,
-        ]);
-        return response()->json([
-            "status" => 200,
-            "Message" => "modification avec succès",
-            "user" => $user
-        ]);
+        if ($user) {
+            // Vérifier si 'nom' est défini avant de l'assigner
+            $nom = $request->has('nom') ? $request->nom : $user->nom;
+            $prenom = $request->has('prenom') ? $request->prenom : $user->prenom;
+            $telephone = $request->has('telephone') ? $request->telephone : $user->telephone;
+            $email = $request->has('email') ? $request->email : $user->email;
+            $specialite = $request->has('specialite') ? $request->specialite : $user->specialite;
+            $password = $request->has('password') ? $request->password : $user->password;
+            $role = $request->has('role') ? $request->role : $user->role;
+
+            $user->update([
+                "nom" => $nom,
+                "prenom" => $prenom,
+                "telephone" => $telephone,
+                "email" => $email,
+                "specialite" => $specialite,
+                "password" => bcrypt($password),
+                "role" => $role,
+            ]);
+            //$user = $user->fresh();
+            return response()->json([
+                "data" => [
+                    "status" => 200,
+                    "message" => "modification avec succès",
+                    "user" => $user
+                ]
+            ]);
+        } else {
+            return response()->json([
+                "data" => [
+                    "status" => 404,
+                    "message" => "Utilisateur non trouvé."
+                ]
+            ], 404);
+        }
     }
 
     /**
@@ -103,12 +165,23 @@ class AdminController extends Controller
     public function destroy(string $id)
     {
         $user = User::find($id);
-        $user->delete();
-        return response()->json(
-            [
-                'message' => 'User supprimé avec succès',
-                'status' => 200
-            ]
-        );
+
+        if ($user && $user->exists()) {
+            $user->delete();
+
+            return response()->json([
+                "data" => [
+                    'message' => 'Utilisateur supprimé avec succès',
+                    'status' => 200
+                ]
+            ]);
+        } else {
+            return response()->json([
+                "data" => [
+                    'message' => 'Utilisateur non trouvé',
+                    'status' => 404
+                ]
+            ], 404);
+        }
     }
 }
